+++ lib/util/file_picker.dart
@@ -2,7 +2,7 @@ import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:localsend_app/gen/strings.g.dart';
-import 'package:localsend_app/provider/selected_files_provider.dart';
+import 'package:localsend_app/provider/selection/selected_sending_files_provider.dart';
 import 'package:localsend_app/widget/dialogs/message_input_dialog.dart';
 import 'package:wechat_assets_picker/wechat_assets_picker.dart';
 
@@ -34,7 +34,7 @@ enum FilePickerOption {
       case FilePickerOption.file:
         final result = await FilePicker.platform.pickFiles(allowMultiple: true);
         if (result != null) {
-          ref.read(selectedFilesProvider.notifier).addFiles(
+          ref.read(selectedSendingFilesProvider.notifier).addFiles(
             files: result.files,
             converter: CrossFileConverters.convertPlatformFile,
           );
@@ -48,7 +48,7 @@ enum FilePickerOption {
           ),
         );
         if (result != null) {
-          ref.read(selectedFilesProvider.notifier).addFiles(
+          ref.read(selectedSendingFilesProvider.notifier).addFiles(
             files: result,
             converter: CrossFileConverters.convertAssetEntity,
           );
@@ -57,7 +57,7 @@ enum FilePickerOption {
       case FilePickerOption.text:
         final result = await showDialog<String>(context: context, builder: (_) => const MessageInputDialog());
         if (result != null) {
-          ref.read(selectedFilesProvider.notifier).addMessage(result);
+          ref.read(selectedSendingFilesProvider.notifier).addMessage(result);
         }
         break;
     }