+++ lib/widget/labeled_checkbox.dart
@@ -1,0 +1,26 @@
+import 'package:flutter/material.dart';
+
+class LabeledCheckbox extends StatelessWidget {
+  final String label;
+  final bool value;
+  final ValueChanged<bool?> onChanged;
+  const LabeledCheckbox({
+    required this.label,
+    required this.value,
+    required this.onChanged,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      children: [
+        Checkbox(
+          value: value,
+          onChanged: onChanged,
+        ),
+        const SizedBox(width: 5),
+        Text(label),
+      ],
+    );
+  }
+}